django/requirements.txt
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary==2.9.2
drf-writable-nested==0.6.3
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-storages==1.12.3
Pillow==8.4.0
boto3==1.20.24
py-eth-sig-utils==0.4.0
djangorestframework-simplejwt==5.0.0
easy-thumbnails==2.8
django-sslserver==0.22
openai==0.15.0
multicolorcaptcha==1.2.0
sentry-sdk==1.9.3
gspread==5.5.0
django-background-tasks==1.2.5
